[CMREC-1420] Pruebas del WS de alta de tasación
Former-commit-id: 9b73a67cbe8d931fb24c06f7f67e692158df196b

+++ proyecto-cajamar-ws/src/main/java/es/pfsgroup/recovery/cajamar/ws/SolicitarTasacionWS.java
@@ -133,14 +133,15 @@ public class SolicitarTasacionWS extends BaseWS implements SolicitarTasacionWSAp
 			for (NMBValoracionesBien val : valoraciones) {
 				if (val.getId() == valoracionActiva.getId() ) 
 					nueva = val;
-					nueva.setCodigoNuita(new Integer(output.getIDTA()));
+				
+					nueva.setCodigoNuita(new Long(output.getIDTA()));
 					nueva.setFechaSolicitudTasacion(new Date());
 					Auditoria auditoria = Auditoria.getNewInstance();
 					nueva.setAuditoria(auditoria);
 					break;
 	        }
 		} else {
-			nueva.setCodigoNuita(new Integer(output.getIDTA()));
+			nueva.setCodigoNuita(new Long(output.getIDTA()));
 			nueva.setBien(bien);
 			Auditoria auditoria = Auditoria.getNewInstance();
 			nueva.setAuditoria(auditoria);
